requirements.txt
PyQt6>=6.4.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
mutagen>=1.45.1
numpy>=1.22.0
matplotlib>=3.5.1
pillow>=9.0.0
pyinstaller>=5.1.0
pytest>=7.0.0
qtawesome>=1.1.1
requests>=2.28.0 